re-org utils funs to new file, cleanup colnamesInt
