+++ inst/tests/nafill.Rraw
@@ -133,10 +133,10 @@ test(4.03, colnamesInt(dt, 1), 1L)
 test(4.04, colnamesInt(dt, c("a","d")), c(1L, 3L))
 test(4.05, colnamesInt(dt, c(1L, 3L)), c(1L, 3L))
 test(4.06, colnamesInt(dt, c(1, 3)), c(1L, 3L))
-test(4.07, colnamesInt(dt, c("a", "e")), error="argument specify non existing column")
-test(4.08, colnamesInt(dt, c(1L, 4L)), error="argument specify non existing column")
-test(4.09, colnamesInt(dt, c(1, 4)), error="argument specify non existing column")
-test(4.10, colnamesInt(dt, c("a", NA)), error="argument specify non existing column")
+test(4.07, colnamesInt(dt, c("a", "e")), error="argument specify non existing column*.*e")
+test(4.08, colnamesInt(dt, c(1L, 4L)), error="argument specify non existing column*.*4")
+test(4.09, colnamesInt(dt, c(1, 4)), error="argument specify non existing column*.*4")
+test(4.10, colnamesInt(dt, c("a", NA)), error="argument specify non existing column*.*NA")
 test(4.11, colnamesInt(dt, c(1L, NA)), error="argument specify non existing column")
 test(4.12, colnamesInt(dt, c(1, NA)), error="argument specify non existing column")
 test(4.13, colnamesInt(dt, c("a","d","a")), error="argument specify duplicated column")
